eeg_emotion/requirements.txt
numpy>=1.24
scipy>=1.10
h5py>=3.8
scikit-learn>=1.3
torch>=2.0
pyyaml>=6.0
openpyxl>=3.1
tqdm>=4.66
